splitted .h and .cxx files into two by classes
