+++ modules/litsov-alex-binarytree/include/BinaryTree.h
@@ -2,34 +2,10 @@
 #ifndef MODULES_LITSOV_ALEX_BINARYTREE_INCLUDE_BINARYTREE_H_
 #define MODULES_LITSOV_ALEX_BINARYTREE_INCLUDE_BINARYTREE_H_
 
+#include "Element.h"
 #include <string>
 using std::string;
 
-class Element {
- public:
-    explicit Element(const int& init_key = 0, const string& init_value = "",
-        Element* const init_parent = 0, Element* const init_right = 0,
-        Element* const init_left = 0);
-    Element(const Element& init_element);
-    ~Element();
-    int GetKey() const;
-    void SetKey(const int& assigned_key);
-    string GetValue() const;
-    void SetValue(const string& assigned_value);
-    void SetRight(Element* const  assigned_right);
-    void SetLeft(Element* const assigned_left);
-    void SetParent(Element* const assigned_parent);
-    Element* GetRight() const;
-    Element* GetLeft() const;
-    Element* GetParent() const;
-
- private:
-    Element* parent_;
-    Element *right_;
-    Element *left_;
-    int key_;
-    string value_;
-};
 class BinaryTree {
  public:
     explicit BinaryTree(Element* const init_root = 0);
@@ -43,9 +19,7 @@ class BinaryTree {
         Element* const start_element) const;
     Element* SearchByKey(const int& key) const;
     Element* SearchByValue(const string& value) const;
-    Element* GetMinByKeyInSubtree(Element* const start_element) const;
     Element* GetMinByKey() const;
-    Element* GetMaxByKeyInSubtree(Element* const start_element) const;
     Element* GetMaxByKey() const;
     Element* GetSuccessor(Element* const specified_element) const;
     Element* GetPredecessor(Element* const specified_element) const;
@@ -57,6 +31,8 @@ class BinaryTree {
  private:
     Element* CopySubtree(Element* subtree_root,
         Element* roots_parent = 0) const;
+    Element* GetMinByKeyInSubtree(Element* const start_element) const;
+    Element* GetMaxByKeyInSubtree(Element* const start_element) const;
     void GetKeysOrderInSubtreeRecurse(Element* start_element,
         string* order) const;
     void GetValuesOrderInSubtreeRecurse(Element* start_element,
